feat [main.js] add modules structure of .js files
