+++ index.html
@@ -232,8 +232,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-<script src="/js/functions.js"></script>
-<script src="/js/main.js"></script>
+<script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
